requirements.txt
streamlit==1.54.0

langchain==1.2.10
langchain-core==1.2.12
langchain-community==0.4.1
langchain-openai==1.1.9
langchain-chroma==1.1.0
langchain-huggingface==1.2.0

chromadb==1.5.0

sentence-transformers==5.2.2
torch==2.2.2
transformers==4.41.2
huggingface-hub==0.23.2

openai==2.20.0
tiktoken==0.12.0

pypdf==4.2.0
python-dotenv==1.2.1
